requirements.txt
pillow
numpy
flask
matplotlib
request

